Safari-609.1.11.1/LayoutTests/js/typedarray-set-same-type-memmove-expected.txt

Tests the code path of typedArray.set that bottoms out in memmove.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(10) is [42,42,43,44,45,46,47,48,49,50,51]
PASS bar(10) is [42,43,44,45,46,47,48,49,50,51,51]
PASS successfullyParsed is true

TEST COMPLETE
